+++ src/main/java/de/hpi/msc/jschneider/math/Intersection.java
@@ -6,6 +6,6 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder @Getter
 public class Intersection
 {
-    private float intersectionDistance;
+    private double intersectionDistance;
     private long subSequenceIndex;
 }
