+++ src/main/java/de/hpi/msc/jschneider/protocol/sequenceSliceDistribution/receiver/SequenceSliceReceiverModel.java
@@ -26,15 +26,15 @@ public class SequenceSliceReceiverModel extends AbstractProtocolParticipantModel
     @Setter @Getter
     private int convolutionSize;
     @Builder.Default @Getter @Setter
-    private float minimumRecord = Float.MAX_VALUE;
+    private double minimumRecord = Double.MAX_VALUE;
     @Builder.Default @Getter @Setter
-    private float maximumRecord = Float.MIN_VALUE;
+    private double maximumRecord = Double.MIN_VALUE;
     @Setter @Getter
     private MatrixInitializer projectionInitializer;
     @Setter @Getter @Builder.Default
-    private float[] unusedRecords = new float[0];
+    private double[] unusedRecords = new double[0];
     @Setter @Getter
-    private List<Float> rawSubSequence;
+    private List<Double> rawSubSequence;
     @NonNull @Getter @Builder.Default
     private SequenceWriter sequenceWriter = BinaryDirectoryWriter.fromDirectory(Paths.get(SystemParameters.getWorkingDirectory().toString(), "sequence-slices/").toFile());
 }
